requirements.txt
streamlit
pandas
scikit-learn
python-docx
PyMuPDF
